test(alert): migrate trigger
